See #12. Moves direction tests into 'test_direction.cpp' and removes Catch2 test
